refactor(artifacts): Move copy-pasted exception to its own class
The exception FailedDownloadException is duplicated in 4 different
classes; pull it out into its own file.
