+++ clouddriver-artifacts/src/main/java/com/netflix/spinnaker/clouddriver/artifacts/bitbucket/BitbucketArtifactCredentials.java
@@ -20,6 +20,7 @@ package com.netflix.spinnaker.clouddriver.artifacts.bitbucket;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.netflix.spinnaker.clouddriver.artifacts.config.ArtifactCredentials;
+import com.netflix.spinnaker.clouddriver.artifacts.exceptions.FailedDownloadException;
 import com.netflix.spinnaker.kork.artifacts.model.Artifact;
 import com.squareup.okhttp.HttpUrl;
 import com.squareup.okhttp.OkHttpClient;
@@ -103,16 +104,6 @@ public class BitbucketArtifactCredentials implements ArtifactCredentials {
       return downloadResponse.body().byteStream();
     } catch (IOException e) {
       throw new FailedDownloadException("Unable to download the contents of artifact " + artifact + ": " + e.getMessage(), e);
-    }
-  }
-
-  public class FailedDownloadException extends IOException {
-    public FailedDownloadException(String message) {
-      super(message);
-    }
-
-    public FailedDownloadException(String message, Throwable cause) {
-      super(message, cause);
     }
   }
 }
